Refactor game logic
This primarily involved lifting the clicked state to the grandparent component, App.jsx, as well as the handling logic on a bird click.
However, to manage these updates with the array shuffling, I mapped id's to the boolean clicked value in it's own object.
On a repeat click, the birdClicked object keys' values are reset to false so the user can play again.