Fix game reset logic
There was some issues with the game resetting and updating the high score when the user clicked all the cards.

The issue lied with the high score setting logic, as it utilized the score variable after the setState function was invoked but before the component was re-rendered.

To fix this problem, I simply added a win parameter to the resetBirdClicked function to simply conditionally branch - if the user won, the high score was set to the # of birds in the list, and if the user lost it was set to the current score.

This may not be the best approach, but it is one that worked.